pyproject.toml
[project]
name = "net-queue"
version = "2.2.0"
description = "Easy & fast network communications using queues"
readme = "README.md"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
authors = [
    { name = "Miguel Ángel Prosper", email = "mprosper@uji.es" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
keywords = ["Communication", "Protocol", "Python"]
urls = { Homepage = "https://github.com/hpca-uji/net-queue", Repository = "https://github.com/hpca-uji/net-queue.git"}

requires-python = ">=3.13"
dependencies = [
    "bidict>=0.23.1",
    "sabctools>=9.3.0",
    "grpcio>=1.74.0",
    "paho-mqtt>=2.1.0"
]

[project.optional-dependencies]
test = [
    "numpy>=1.19.4"
]

[build-system]
requires = [
    "setuptools>=68.2.0"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "net_queue",
    "net_queue.*"
]